Move FormFooter to its own component to simplify form.js
