+++ src/applications/user-testing/new-conditions/config/form.js
@@ -1,7 +1,6 @@
-import React from 'react';
 import { VA_FORM_IDS } from 'platform/forms/constants';
-import CallVBACenter from 'platform/static-data/CallVBACenter';
 
+import FormFooter from '../components/FormFooter';
 import { SUBTITLE, TITLE } from '../constants';
 import ConfirmationPage from '../containers/ConfirmationPage';
 import IntroductionPage from '../containers/IntroductionPage';
@@ -13,23 +12,6 @@ import followUpCause from '../pages/followUpCause';
 import followUpCauseDescription from '../pages/followUpCauseDescription';
 import followUpDate from '../pages/followUpDate';
 import followUpIntro from '../pages/followUpIntro';
-
-const FormFooter = () => (
-  <div className="row vads-u-margin-bottom--2">
-    <div className="usa-width-two-thirds medium-8 columns">
-      <va-need-help>
-        <div slot="content">
-          <div>
-            <p className="help-talk">
-              For help filling out this form, or if the form isn’t working
-              right, <CallVBACenter />
-            </p>
-          </div>
-        </div>
-      </va-need-help>
-    </div>
-  </div>
-);
 
 /** @type {FormConfig} */
 const formConfig = {
